+++ saleor/graphql/order/mutations/order_note_add.py
@@ -6,7 +6,7 @@ from ....permission.enums import OrderPermissions
 from ...app.dataloaders import get_app_promise
 from ...core import ResolveInfo
 from ...core.descriptions import (
-    ADDED_IN_314,
+    ADDED_IN_315,
     DEPRECATED_IN_3X_INPUT,
     DEPRECATED_IN_3X_MUTATION,
     PREVIEW_FEATURE,
@@ -32,7 +32,7 @@ class OrderNoteAdd(OrderNoteCommon):
         )
 
     class Meta:
-        description = "Adds note to the order." + ADDED_IN_314 + PREVIEW_FEATURE
+        description = "Adds note to the order." + ADDED_IN_315 + PREVIEW_FEATURE
         doc_category = DOC_CATEGORY_ORDERS
         permissions = (OrderPermissions.MANAGE_ORDERS,)
         error_type_class = OrderNoteError