+++ saleor/graphql/order/mutations/order_note_update.py
@@ -5,7 +5,7 @@ from ....order import OrderEvents, events, models
 from ....permission.enums import OrderPermissions
 from ...app.dataloaders import get_app_promise
 from ...core import ResolveInfo
-from ...core.descriptions import ADDED_IN_314, PREVIEW_FEATURE
+from ...core.descriptions import ADDED_IN_315, PREVIEW_FEATURE
 from ...core.doc_category import DOC_CATEGORY_ORDERS
 from ...core.types.common import OrderNoteError
 from ...plugins.dataloaders import get_plugin_manager_promise
@@ -26,7 +26,7 @@ class OrderNoteUpdate(OrderNoteCommon):
         )
 
     class Meta:
-        description = "Updates note of an order." + ADDED_IN_314 + PREVIEW_FEATURE
+        description = "Updates note of an order." + ADDED_IN_315 + PREVIEW_FEATURE
         doc_category = DOC_CATEGORY_ORDERS
         permissions = (OrderPermissions.MANAGE_ORDERS,)
         error_type_class = OrderNoteError